api/requirements.txt
anyio==3.7.0
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
fastapi==0.96.0
h11==0.14.0
idna==3.4
motor==3.1.2
pydantic==1.10.8
PyJWT==2.7.0
pymongo==4.3.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
